shared以下からv2以前の実装の依存を排除
