+++ src/process/onetime/attack/drafting_room_process.ts
@@ -2,7 +2,6 @@ import { PrimitiveLogger } from "os/infrastructure/primitive_logger"
 import { Procedural } from "process/procedural"
 import { Process, ProcessId } from "process/process"
 import { coloredText, profileLink, roomLink } from "utility/log"
-import { RoomCoordinate, RoomCoordinateState } from "shared/utility/room_name"
 import { Timestamp } from "shared/utility/timestamp"
 import { ProcessDecoder } from "process/process_decoder"
 import { ProcessState } from "process/process_state"
@@ -18,6 +17,7 @@ import { Invader } from "game/invader"
 import { GameMap } from "game/game_map"
 import { LeagueOfAutomatedNations } from "_old/loanUserList"
 import type { RoomName } from "shared/utility/room_name_types"
+import { RoomCoordinate, RoomCoordinateState } from "utility/room_coordinate"
 
 ProcessDecoder.register("DraftingRoomProcess", state => {
   return DraftingRoomProcess.decode(state as DraftingRoomProcessState)
