+++ src/process/temporary/season_1244215_generic_dismantle_process.ts
@@ -1,6 +1,5 @@
 import { Procedural } from "process/procedural"
 import { Process, ProcessId } from "process/process"
-import { isRoomName } from "shared/utility/room_name"
 import { roomLink } from "utility/log"
 import { ProcessState } from "process/process_state"
 import { CreepRole } from "prototype/creep_role"
@@ -23,6 +22,7 @@ import { FleeFromAttackerTask } from "v5_object_task/creep_task/combined_task/fl
 import { MoveToTask } from "v5_object_task/creep_task/meta_task/move_to_task"
 import { CreepTask } from "v5_object_task/creep_task/creep_task"
 import type { RoomName } from "shared/utility/room_name_types"
+import { isRoomName } from "utility/room_coordinate"
 
 ProcessDecoder.register("Season1244215GenericDismantleProcess", state => {
   return Season1244215GenericDismantleProcess.decode(state as Season1244215GenericDismantleProcessState)
