+++ src/shared/utility/position.ts
@@ -1,0 +1,4 @@
+export type Position = {
+  readonly x: number
+  readonly y: number
+}
