+++ src/shared/utility/system_info.ts
@@ -22,7 +22,7 @@ export const SystemInfo = {
     name: "AntOS",
   },
   application: {
-    version: "9.0.23",
+    version: "9.0.24",
     name: "Declarative AI",
   },
 }
