+++ src/v5_task/bootstrap_room/upgrade_to_rcl3_task.ts
@@ -1,5 +1,4 @@
 import { ProblemFinder } from "v5_problem/problem_finder"
-import { roomTypeOf } from "shared/utility/room_name"
 import { ChildTaskExecutionResults, Task, TaskIdentifier, TaskStatus } from "v5_task/task"
 import { OwnedRoomObjects } from "world_info/room_info"
 import { GeneralCreepWorkerTask, GeneralCreepWorkerTaskCreepRequest, GeneralCreepWorkerTaskState } from "v5_task/general/general_creep_worker_task"
@@ -29,6 +28,7 @@ import { OwnedRoomResource } from "room_resource/room_resource/owned_room_resour
 import { PrimitiveLogger } from "os/infrastructure/primitive_logger"
 import { roomLink } from "utility/log"
 import type { RoomName } from "shared/utility/room_name_types"
+import { roomTypeOf } from "utility/room_coordinate"
 
 const minimumNumberOfCreeps = 6
 const defaultNumberOfCreeps = 10
