+++ src/v5_task/remote_room_keeper/remote_room_worker_task.ts
@@ -1,4 +1,3 @@
-import { roomTypeOf } from "shared/utility/room_name"
 import { EnergySourceTask } from "v5_task/hauler/owned_room_energy_source_task"
 import { Task, TaskIdentifier, TaskStatus } from "v5_task/task"
 import { TaskState } from "v5_task/task_state"
@@ -8,6 +7,7 @@ import { RemoteRoomReserveTask } from "./remote_room_reserve_task"
 import { ProblemFinder } from "v5_problem/problem_finder"
 import { roomLink } from "utility/log"
 import type { RoomName } from "shared/utility/room_name_types"
+import { roomTypeOf } from "utility/room_coordinate"
 
 export interface RemoteRoomWorkerTaskState extends TaskState {
   /** room name */
