+++ src/v5_task/scout/owned_room_scout_task.ts
@@ -1,8 +1,8 @@
 import { ProblemFinder } from "v5_problem/problem_finder"
-import { roomTypeOf } from "shared/utility/room_name"
 import { Task, TaskIdentifier, TaskStatus } from "v5_task/task"
 import { TaskState } from "v5_task/task_state"
 import type { RoomName } from "shared/utility/room_name_types"
+import { roomTypeOf } from "utility/room_coordinate"
 
 export interface OwnedRoomScoutTaskState extends TaskState {
   /** room name */
