+++ src/v5_task/worker/primitive_worker_task.ts
@@ -1,4 +1,3 @@
-import { roomTypeOf } from "shared/utility/room_name"
 import { Task, TaskIdentifier, TaskStatus } from "v5_task/task"
 import { OwnedRoomObjects } from "world_info/room_info"
 import { CreepRole, hasNecessaryRoles } from "prototype/creep_role"
@@ -26,6 +25,7 @@ import { MoveToRoomTask } from "v5_object_task/creep_task/meta_task/move_to_room
 import { RoomPositionFilteringOptions } from "prototype/room_position"
 import { NormalRoomInfo } from "room_resource/room_info"
 import type { RoomName } from "shared/utility/room_name_types"
+import { roomTypeOf } from "utility/room_coordinate"
 
 function moveToOptions(): MoveToTargetTaskOptions {
   return {
